+++ app/scripts/collections/paginated.js
@@ -3,7 +3,8 @@ import AppCollection from 'scripts/collections/app';
 export default class PaginatedCollection extends AppCollection {
   parse(response) {
     const pagination = response.meta.pagination;
-    const pagesTotal = Math.ceil(pagination.total/pagination.per_page);
+    const pagesTotal = Math.ceil(pagination.total / pagination.per_page);
+    const collectionKey = _.findKey(response, (value) => Array.isArray(value));
 
     this.pagination = {
       pagesTotal,
@@ -12,7 +13,7 @@ export default class PaginatedCollection extends AppCollection {
       total: pagination.total
     };
 
-    return super.parse(response[this.rootKey]);
+    return super.parse(response[collectionKey]);
   }
 
   fetchByPage(page = 1) {